Recover records in Freelist leaf pages (#9)
As far as I can tell, sqlite turns existing B-tree pages into Freelist leaf pages wholesale, i.e. without changing anything in the page itself (to be confirmed, though).
If correct, that means it should still be possible to parse a Freelist leaf page that was formerly a B-tree table leaf page as if it was still a table leaf page.

This works for at least some Freelist leaves, but the next problem is that the ptrmap entry for a recovered Freelist page points to its Freelist trunk page rather than to its former table interior page. Consequently, a new mechanism is required through which Freelist leaf pages can be matched to tables. I'm using a pretty naive signature system that relies on the internal Python types used for the values of fields in a page's records.